# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  aroma-hub_internal_application_dto.CreateOrderRequest:
    properties:
      address:
        type: string
      contactType:
        allOf:
        - $ref: '#/definitions/aroma-hub_internal_models.ContactType'
        enum:
        - telegram
        - phone
      fullName:
        type: string
      paymentMethod:
        allOf:
        - $ref: '#/definitions/aroma-hub_internal_models.PaymentMethod'
        enum:
        - IBAN
        - сash_on_delivery
      phoneNumber:
        type: string
      productItems:
        items:
          $ref: '#/definitions/aroma-hub_internal_application_dto.ProductOrder'
        type: array
      promoCode:
        type: string
    required:
    - address
    - contactType
    - fullName
    - paymentMethod
    - phoneNumber
    - productItems
    type: object
...
  aroma-hub_internal_application_dto.Order:
    properties:
      address:
        type: string
      amountToPay:
        type: integer
      contactType:
        $ref: '#/definitions/aroma-hub_internal_models.ContactType'
      createdAt:
        type: string
      fullName:
        type: string
      id:
        type: string
      paymentMethod:
        $ref: '#/definitions/aroma-hub_internal_models.PaymentMethod'
      phoneNumber:
        type: string
      products:
        items:
          $ref: '#/definitions/aroma-hub_internal_application_dto.ProductOrder'
        type: array
      status:
        $ref: '#/definitions/aroma-hub_internal_models.OrderStatus'
      updatedAt:
        type: string
    type: object
  aroma-hub_internal_application_dto.OrderResponse:
    properties:
      count:
        type: integer
      orders:
        items:
          $ref: '#/definitions/aroma-hub_internal_application_dto.Order'
        type: array
    type: object
  aroma-hub_internal_application_dto.ProductOrder:
    properties:
      brand:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: integer
      quantity:
        type: integer
      volume:
        type: integer
    required:
    - brand
    - id
    - name
    - price
    - quantity
    - volume
    type: object
  aroma-hub_internal_application_dto.UpdateOrderRequest:
    properties:
      address:
        type: string
      fullName:
        type: string
      id:
        type: string
      paymentMethod:
        $ref: '#/definitions/aroma-hub_internal_models.PaymentMethod'
      phoneNumber:
        type: string
      status:
        $ref: '#/definitions/aroma-hub_internal_models.OrderStatus'
    required:
    - id
    type: object
...
  aroma-hub_internal_models.ContactType:
    enum:
    - telegram
    - phone
    type: string
    x-enum-varnames:
    - ContactTypeTelegram
    - ContactTypePhone
  aroma-hub_internal_models.OrderStatus:
    enum:
    - pending
    - processing
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusProcessing
    - OrderStatusCompleted
    - OrderStatusCancelled
  aroma-hub_internal_models.PaymentMethod:
    enum:
    - IBAN
    - сash_on_delivery
    type: string
    x-enum-varnames:
    - PaymentMethodIBAN
    - PaymentMethodCashOnDelivery
...
  errx.Code:
    enum:
    - CONFLICT
    - INTERNAL
    - NOT_FOUND
    - BAD_REQUEST
    - ALREADY_EXISTS
    - UNAUTHORIZED
    - FORBIDDEN
    - TIMEOUT
    - VALIDATION
    type: string
    x-enum-comments:
      AlreadyExists: Resource already exists
      BadRequest: Invalid input or parameters
      Conflict: Resource conflicts with existing data
      Forbidden: Permission denied
      Internal: Internal server or system errors
      NotFound: Resource not found
      Timeout: Operation timed out
      Unauthorized: Authentication required
      Validation: Input validation failed
    x-enum-varnames:
    - Conflict
    - Internal
    - NotFound
    - BadRequest
    - AlreadyExists
    - Unauthorized
    - Forbidden
    - Timeout
    - Validation
  errx.Error:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errx.Code'
        description: Error classification code
      err:
        description: Original error (if any)
      message:
        description: User-friendly error message
        type: string
    type: object
info:
  contact: {}
  description: dAPI documentation.
  title: Aroma-Hub API
  version: "1.0"
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      description: Get a list of orders with optional filtering
      parameters:
      - description: Order ID
        in: query
        name: id
        type: string
      - description: User ID
        in: query
        name: userId
        type: string
      - description: Payment method (IBAN, сash_on_delivery)
        in: query
        name: paymentMethod
        type: string
      - description: Contact type (telegram, phone)
        in: query
        name: contactType
        type: string
      - description: Order status (pending, processing, completed, cancelled)
        in: query
        name: status
        type: string
      - description: 'Start date for filtering (format: YYYY-MM-DD)'
        in: query
        name: fromDate
        type: string
      - description: 'End date for filtering (format: YYYY-MM-DD)'
        in: query
        name: toDate
        type: string
      - description: 'Number of items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of orders
          schema:
            $ref: '#/definitions/aroma-hub_internal_application_dto.OrderResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/errx.Error'
        "404":
          description: No orders found
          schema:
            $ref: '#/definitions/errx.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errx.Error'
      summary: List orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order
      parameters:
      - description: Order information
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/aroma-hub_internal_application_dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/errx.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errx.Error'
      summary: Create order
      tags:
      - orders
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/errx.Error'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/errx.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errx.Error'
      summary: Delete order
      tags:
      - orders
    put:
      consumes:
      - application/json
      description: Update an existing order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Order information
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/aroma-hub_internal_application_dto.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/errx.Error'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/errx.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errx.Error'
      summary: Update order
      tags:
      - orders
  /orders/{id}/cancel:
    put:
      consumes:
      - application/json
      description: Cancel an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order cancelled
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/errx.Error'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/errx.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errx.Error'
      summary: Cancel order
      tags:
      - orders
...
swagger: "2.0"